+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - `Innmind\OperatingSystem\Filesystem` is now a final class
 - `Innmind\OperatingSystem\CurrentProcess\Signals` is now a final class
 - `Innmind\OperatingSystem\CurrentProcess` is now a final class
+- `Innmind\OperatingSystem\OperatingSystem::map()` callable must now return a `Config`
 
 ### Fixed